Mark an expression as a meta-expression

Mark an expression as a meta-expression

metaExpr(expr, env = parent.frame(), quoted = FALSE,
  localize = "auto", bindToReturn = FALSE)

Arguments

expr
An expression (quoted or unquoted).

env
An environment.

quoted
Is the expression quoted? This is useful when you want to use an expression that is stored in a variable; to do so, it must be quoted with quote().

localize
Whether or not to wrap the returned expression in local(). The default, "auto", only wraps expressions with a top-level return() statement (i.e., return statements in anonymized functions are ignored).

bindToReturn
For non-localized expressions, should an assignment of a meta expression be applied to the last child of the top-level { call?

See also