Produce a zip bundle of code and results

Produce a zip bundle of code and results

buildScriptBundle(
  code = NULL,
  output_zip_path,
  script_name = "script.R",
  include_files = list(),
  render = TRUE,
  render_args = list()
)

buildRmdBundle(
  report_template,
  output_zip_path,
  vars = list(),
  include_files = list(),
  render = TRUE,
  render_args = list()
)

Arguments

code

A language object.

output_zip_path

A filename for the resulting zip bundle.

script_name

A name for the R script in the zip bundle.

include_files

A named list consisting of additional files that should be included in the zip bundle. The element names indicate the destination path within the bundle, specified as a relative path; the element values indicate the path to the actual file currently on disk, specified as either a relative or absolute path.

render

Whether or not to call rmarkdown::render() on the R script.

render_args

Arguments to provide to rmarkdown::render().

report_template

Filename of an Rmd template to be expanded by knitr::knit_expand().

vars

A named list of variables passed along to ... in knitr::knit_expand().

Value

The path to a generated file.

See also